Create EHYHB.ino
My previous attempt with header .h file was not good. In my second attempt, I have simply moved one function to a separate .ino file. The idea is that there will be several model-specific .ino files with convertPacket() function. But user will only keep one of these .ino files in the folder. 

I do not know about better solution (I know nothing about C++ programing and I do not know how to build libraries for Arduino). So suggestions for better solution welcomed. 
Is it possible to build some kind of library which can be used both in Arduino and in Raspberry (Python)?