LogicalFormat: Add EJHA04AAV3 protocol
The EJHA04AAV3 is an hybrid air-to-water heat pump in cascade
with gas condensing technology. It does not support cooling.

The protocol is pretty similar to EHYHBX08AAV3 and EHVX08S26CB9W.
So far the only difference seems to be that it has less sensors
(flow, energy produced) than the existing ones.

Packets 10h-16h and b8h have been observed during
passive sniffing on the P1P2 bus. Packets 00f031h-00f038h have
been observed with an external controller on the bus.

Signed-off-by: Patrick Rudolph <rudolphpatrick05@gmail.com>